requirements.txt
pdf2image==1.5.4
Pillow==6.0.0
pkg-resources==0.0.0
pytesseract==0.2.6
